service provider finder

+++ lib/application/classes/user/user_by_distance.dart
@@ -5,14 +5,14 @@ class UserByDistance {
   String longitude;
   String address;
   String distance;
-  int rating;
+  double rating;
   UserByDistance.fromJson(Map<String, dynamic> jsonMap)
       : id = jsonMap['identifier'].toString(),
         userName = jsonMap['username'],
         latitude = jsonMap['latitude'].toString(),
         longitude = jsonMap['longitude'].toString(),
         address = jsonMap['address'],
-        distance = double.parse((jsonMap['distance']).toStringAsFixed(2)).toString(),
-        rating = jsonMap['rate'];
-
+        distance =
+            double.parse((jsonMap['distance']).toStringAsFixed(2)).toString(),
+        rating = jsonMap['rate'].toDouble();
 }
